Model tests and indexes

+++ controllers/forms/city.go
@@ -2,14 +2,14 @@ package forms
 
 //City : Represents the form data for creating a city
 type City struct {
-	Name      string `json:"name" binding:"required"`
-	Latitude  int    `json:"latitude" binding:"required"`
-	Longitude int    `json:"longitude" binding:"required"`
+	Name      string  `json:"name" binding:"required"`
+	Latitude  float64 `json:"latitude" binding:"required"`
+	Longitude float64 `json:"longitude" binding:"required"`
 }
 
 //UpdateCity : Represents the form data for creating a city
 type UpdateCity struct {
-	Name      string `json:"name"`
-	Latitude  int    `json:"latitude"`
-	Longitude int    `json:"longitude"`
+	Name      string  `json:"name"`
+	Latitude  float64 `json:"latitude"`
+	Longitude float64 `json:"longitude"`
 }
